Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo[srv]==3.11
motor
passlib[bcrypt]
python-jose
spacy
PyMuPDF
scikit-learn
cloudinary
requests
python-multipart